Create bool to string converstion
The header stdbool. h in the C Standard Library for the C programming language contains four macros for a Boolean data type. This header was introduced in C99.
Using the system header file stdbool. h allows you to use bool as a Boolean data type. true evaluates to 1 and false evaluates to 0